requirements.txt
alembic==1.14.0
bcrypt==4.2.1
blinker==1.8.2
certifi==2024.8.30
cffi==1.17.1
chardet==3.0.4
click==8.1.7
colorama==0.4.6
cryptography==44.0.0
Deprecated==1.2.15
distlib==0.3.8
filelock==3.15.4
Flask==2.3.2
Flask-Bcrypt==1.0.1
Flask-Cors==3.0.8
Flask-Limiter==2.8.1
Flask-Login==0.6.3
Flask-Migrate==4.1.0
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.1.1
greenlet==3.1.1
idna==2.6
itsdangerous==2.2.0
Jinja2==3.1.4
limits==3.13.0
Mako==1.3.8
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
ordered-set==4.1.0
packaging==24.2
platformdirs==4.2.2
pure-eval==0.2.2
pycodestyle==2.6.0
pycparser==2.22
Pygments==2.18.0
pyotp==2.9.0
python-dotenv==1.0.1
requests==2.18.4
rich==12.6.0  # Updated to a version compatible with Flask-Limiter 2.8.1
six==1.16.0
SQLAlchemy==2.0.36
traitlets==5.14.3
typing_extensions==4.12.2
urllib3==1.22
virtualenv==20.26.3
wcwidth==0.2.13
Werkzeug==2.3.3
wrapt==1.17.1
WTForms==3.0.1
zxcvbn-python==4.4.24